// frontend/src/app/api/customGraphql.service.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: any }> = { [K in keyof T]: T[K] };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
    ID: string;
    String: string;
    Boolean: boolean;
    Int: number;
    Float: number;
    /** The `Upload` scalar type represents a file upload. */
    Upload: any;
};


export type Query = {
    __typename?: 'Query';
    queryUser?: Maybe<User>;
    queryUserStockWatchlists?: Maybe<Array<Maybe<StockWatchlist>>>;
};


export type QueryQueryUserArgs = {
    uid: Scalars['String'];
};


export type QueryQueryUserStockWatchlistsArgs = {
    uid: Scalars['String'];
};

export type Mutation = {
    __typename?: 'Mutation';
    updateUserData?: Maybe<User>;
    createStockWatchlist?: Maybe<StockWatchlist>;
    renameStockWatchlist?: Maybe<Scalars['Boolean']>;
    deleteWatchlist?: Maybe<Scalars['Boolean']>;
    addStockIntoStockWatchlist?: Maybe<StockDetails>;
    removeStockFromStockWatchlist?: Maybe<Scalars['Boolean']>;
};
// ...
export type User = {
    __typename?: 'User';
    uid: Scalars['ID'];
    displayName?: Maybe<Scalars['String']>;
    email?: Maybe<Scalars['String']>;
    photoURL?: Maybe<Scalars['String']>;
    providerId?: Maybe<Scalars['String']>;
    nickname?: Maybe<Scalars['String']>;
    locale?: Maybe<Scalars['String']>;
    stockWatchlist: Array<Maybe<StockWatchlist>>;
};

export type UserInput = {
    uid: Scalars['ID'];
    displayName?: Maybe<Scalars['String']>;
    email?: Maybe<Scalars['String']>;
    photoURL?: Maybe<Scalars['String']>;
    providerId?: Maybe<Scalars['String']>;
    nickname?: Maybe<Scalars['String']>;
    locale?: Maybe<Scalars['String']>;
};

export type StockWatchlist = {
    __typename?: 'StockWatchlist';
    id: Scalars['String'];
    name: Scalars['String'];
    timestamp: Scalars['Float'];
    userId: Scalars['String'];
    stocks: Array<Maybe<Scalars['String']>>;
    stocksDetails: Array<Maybe<StockDetails>>;
};

export type StockDetails = {
    __typename?: 'StockDetails';
    id: Scalars['String'];
    basicInfo?: Maybe<BasicInfo>;
    chartInfo?: Maybe<ChartInfo>;
    dividend?: Maybe<Dividend>;
    analysis?: Maybe<Analysis>;
    balanceSheet?: Maybe<BalanceSheet>;
    cashFlow?: Maybe<CashFlow>;
    overview?: Maybe<OverView>;
    perShare?: Maybe<PerShare>;
    valuation?: Maybe<Valuation>;
    financialStrength?: Maybe<FinancialStrength>;
    financialReports?: Maybe<Array<Maybe<FinancialReportNames>>>;
    financialStrengthRatio?: Maybe<FinancialStrengthRatio>;
    incomeStatement?: Maybe<IncomeStatement>;
    recommendation?: Maybe<Array<Maybe<Recommendations>>>;
    stockNewsSnippets?: Maybe<Array<Maybe<StockArticle>>>;
};
// ...
export type FinancialReportNames = {
    __typename?: 'FinancialReportNames';
    collection?: Maybe<Scalars['String']>;
    name?: Maybe<Scalars['String']>;
};
// ...
export type CashFlow = {
    __typename?: 'CashFlow';
    Capital_Expenditure?: Maybe<Array<Maybe<Scalars['Float']>>>;
    End_Cash_Position?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Financing_Cash_Flow?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Free_Cash_Flow?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Investing_Cash_Flow?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Issuance_of_Capital_Stock?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Issuance_of_Debt?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Operating_Cash_Flow?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Repayment_of_Debt?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Repurchase_of_Capital_Stock?: Maybe<Array<Maybe<Scalars['Float']>>>;
    date?: Maybe<Array<Maybe<Scalars['String']>>>;
};

export type BalanceSheet = {
    __typename?: 'BalanceSheet';
    Capital_Lease_Obligations?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Common_Stock_Equity?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Invested_Capital?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Net_Debt?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Net_Tangible_Assets?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Ordinary_Shares_Float?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Share_Issued?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Tangible_Book_Value?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Total_Assets?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Total_Capitalization?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Total_Debt?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Total_Equity_Gross_Minority_Interest?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Total_Liabilities_Net_Minority_Interest?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Working_Capital?: Maybe<Array<Maybe<Scalars['Float']>>>;
    date?: Maybe<Array<Maybe<Scalars['String']>>>;
};
// ...
export type Analysis = {
    __typename?: 'Analysis';
    growthEstimatesPercent?: Maybe<GrowthEstimatesPercent>;
    revenueEstimate?: Maybe<RevenueEstimate>;
    earnings?: Maybe<Earnings>;
};

export type ChartInfo = {
    __typename?: 'ChartInfo';
    assetsToDebtInfo?: Maybe<AssetsToDebtInfo>;
    equityToAssets?: Maybe<EquityToAssets>;
    effectiveness?: Maybe<Effectiveness>;
    margin?: Maybe<Margin>;
    otherGrowthInformation?: Maybe<OtherGrowthInformation>;
    profitMargin?: Maybe<ProfitMargin>;
};
// ...
export type Dividend = {
    __typename?: 'Dividend';
    dividendDate?: Maybe<Scalars['String']>;
    dividendPerShare5Y?: Maybe<Scalars['Float']>;
    dividendPerShareAnnual?: Maybe<Scalars['Float']>;
    exDividendDate?: Maybe<Scalars['String']>;
    fiveYearAverageDividendYield?: Maybe<Scalars['String']>;
    forwardAnnualDividendRate?: Maybe<Scalars['String']>;
    forwardAnnualDividendYield?: Maybe<Scalars['String']>;
    payoutRatio?: Maybe<Scalars['String']>;
    trailingAnnualDividendRate?: Maybe<Scalars['String']>;
    trailingAnnualDividendYield?: Maybe<Scalars['String']>;
};

export type BasicInfo = {
    __typename?: 'BasicInfo';
    address1?: Maybe<Scalars['String']>;
    city?: Maybe<Scalars['String']>;
    country?: Maybe<Scalars['String']>;
    exchangeTimezoneName?: Maybe<Scalars['String']>;
    fullTimeEmployees?: Maybe<Scalars['Float']>;
    industry?: Maybe<Scalars['String']>;
    sector?: Maybe<Scalars['String']>;
    shortName?: Maybe<Scalars['String']>;
    state?: Maybe<Scalars['String']>;
    website?: Maybe<Scalars['String']>;
    zip?: Maybe<Scalars['String']>;
    netIncomeEmployeeAnnual?: Maybe<Scalars['Float']>;
    revenueEmployeeAnnual?: Maybe<Scalars['Float']>;
    sharesOutstanding?: Maybe<Scalars['Float']>;
    summary?: Maybe<Scalars['String']>;
    symbol?: Maybe<Scalars['String']>;
    logoUrl?: Maybe<Scalars['String']>;
};
// ...
export type StockWatchlistIdentifier = {
    userId: Scalars['String'];
    id?: Maybe<Scalars['String']>;
    additionalData?: Maybe<Scalars['String']>;
};
// ...
export type StockMainDetailsFragment = (
    { __typename?: 'StockDetails' }
    & Pick<StockDetails, 'id'>
    & {
    overview?: Maybe<(
        { __typename?: 'OverView' }
        & Pick<OverView, 'symbol' | 'currentPrice' | 'weekHigh52' | 'weekLow52' | 'previousClose' | 'earningsDate' | 'exDividendDate' | 'forwardDividendAndYield' | 'targetEst1y'>
        )>, basicInfo?: Maybe<(
        { __typename?: 'BasicInfo' }
        & Pick<BasicInfo, 'sector' | 'industry' | 'logoUrl' | 'shortName' | 'website'>
        )>
}
    );

export type StockWatchlistInformationFragment = (
    { __typename?: 'StockWatchlist' }
    & Pick<StockWatchlist, 'id' | 'name' | 'timestamp' | 'stocks'>
    & {
    stocksDetails: Array<Maybe<(
        { __typename?: 'StockDetails' }
        & StockMainDetailsFragment
        )>>
}
    );

export type QueryUserStockWatchlistsQueryVariables = Exact<{
    uid: Scalars['String'];
}>;


export type QueryUserStockWatchlistsQuery = (
    { __typename?: 'Query' }
    & {
    queryUserStockWatchlists?: Maybe<Array<Maybe<(
        { __typename?: 'StockWatchlist' }
        & StockWatchlistInformationFragment
        )>>>
}
    );

export type CreateStockWatchlistMutationVariables = Exact<{
    identifier: StockWatchlistIdentifier;
}>;


export type CreateStockWatchlistMutation = (
    { __typename?: 'Mutation' }
    & {
    createStockWatchlist?: Maybe<(
        { __typename?: 'StockWatchlist' }
        & StockWatchlistInformationFragment
        )>
}
    );

export type AddStockIntoWatchlistMutationVariables = Exact<{
    identifier: StockWatchlistIdentifier;
}>;


export type AddStockIntoWatchlistMutation = (
    { __typename?: 'Mutation' }
    & {
    addStockIntoStockWatchlist?: Maybe<(
        { __typename?: 'StockDetails' }
        & StockMainDetailsFragment
        )>
}
    );

export type RemoveStockFromWatchlistMutationVariables = Exact<{
    identifier: StockWatchlistIdentifier;
}>;


export type RemoveStockFromWatchlistMutation = (
    { __typename?: 'Mutation' }
    & Pick<Mutation, 'removeStockFromStockWatchlist'>
    );

export type DeleteUserWatchlistMutationVariables = Exact<{
    identifier: StockWatchlistIdentifier;
}>;


export type DeleteUserWatchlistMutation = (
    { __typename?: 'Mutation' }
    & Pick<Mutation, 'deleteWatchlist'>
    );

export type RenameStockWatchlistMutationVariables = Exact<{
    identifier: StockWatchlistIdentifier;
}>;


export type RenameStockWatchlistMutation = (
    { __typename?: 'Mutation' }
    & Pick<Mutation, 'renameStockWatchlist'>
    );

export const StockMainDetailsFragmentDoc = gql`
    fragment StockMainDetails on StockDetails {
        id
        overview {
            symbol
            currentPrice
            weekHigh52
            weekLow52
            previousClose
            earningsDate
            exDividendDate
            forwardDividendAndYield
            targetEst1y
        }
        basicInfo {
            sector
            industry
            logoUrl
            shortName
            website
        }
    }
`;
export const StockWatchlistInformationFragmentDoc = gql`
    fragment StockWatchlistInformation on StockWatchlist {
        id
        name
        timestamp
        stocks
        stocksDetails {
            ...StockMainDetails
        }
    }
${StockMainDetailsFragmentDoc}`;
// ...
export const QueryUserStockWatchlistsDocument = gql`
    query queryUserStockWatchlists($uid: String!) {
        queryUserStockWatchlists(uid: $uid) {
            ...StockWatchlistInformation
        }
    }
${StockWatchlistInformationFragmentDoc}`;

@Injectable({
    providedIn: 'root'
})
export class QueryUserStockWatchlistsGQL extends Apollo.Query<QueryUserStockWatchlistsQuery, QueryUserStockWatchlistsQueryVariables> {
    document = QueryUserStockWatchlistsDocument;

}

export const CreateStockWatchlistDocument = gql`
    mutation CreateStockWatchlist($identifier: StockWatchlistIdentifier!) {
        createStockWatchlist(identifier: $identifier) {
            ...StockWatchlistInformation
        }
    }
${StockWatchlistInformationFragmentDoc}`;

@Injectable({
    providedIn: 'root'
})
export class CreateStockWatchlistGQL extends Apollo.Mutation<CreateStockWatchlistMutation, CreateStockWatchlistMutationVariables> {
    document = CreateStockWatchlistDocument;

}

export const AddStockIntoWatchlistDocument = gql`
    mutation AddStockIntoWatchlist($identifier: StockWatchlistIdentifier!) {
        addStockIntoStockWatchlist(identifier: $identifier) {
            ...StockMainDetails
        }
    }
${StockMainDetailsFragmentDoc}`;

@Injectable({
    providedIn: 'root'
})
export class AddStockIntoWatchlistGQL extends Apollo.Mutation<AddStockIntoWatchlistMutation, AddStockIntoWatchlistMutationVariables> {
    document = AddStockIntoWatchlistDocument;

}

export const RemoveStockFromWatchlistDocument = gql`
    mutation RemoveStockFromWatchlist($identifier: StockWatchlistIdentifier!) {
        removeStockFromStockWatchlist(identifier: $identifier)
    }
`;

@Injectable({
    providedIn: 'root'
})
export class RemoveStockFromWatchlistGQL extends Apollo.Mutation<RemoveStockFromWatchlistMutation, RemoveStockFromWatchlistMutationVariables> {
    document = RemoveStockFromWatchlistDocument;

}

export const DeleteUserWatchlistDocument = gql`
    mutation DeleteUserWatchlist($identifier: StockWatchlistIdentifier!) {
        deleteWatchlist(identifier: $identifier)
    }
`;

@Injectable({
    providedIn: 'root'
})
export class DeleteUserWatchlistGQL extends Apollo.Mutation<DeleteUserWatchlistMutation, DeleteUserWatchlistMutationVariables> {
    document = DeleteUserWatchlistDocument;

}

export const RenameStockWatchlistDocument = gql`
    mutation RenameStockWatchlist($identifier: StockWatchlistIdentifier!) {
        renameStockWatchlist(identifier: $identifier)
    }
`;

@Injectable({
    providedIn: 'root'
})
export class RenameStockWatchlistGQL extends Apollo.Mutation<RenameStockWatchlistMutation, RenameStockWatchlistMutationVariables> {
    document = RenameStockWatchlistDocument;

}